requirements.txt
account==0.1.0
alembic==1.10.2
amqp==2.6.1
anyio==3.6.2
asgiref==3.6.0
attrs==22.2.0
bcrypt==4.0.1
billiard==3.6.4.0
cachetools==5.3.0
celery==4.4.7
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==7.1.2
cryptography==39.0.2
cssselect==1.2.0
cssutils==2.6.0
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.53.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
emails==0.5.15
fastapi==0.70.1
gunicorn==20.1.0
h11==0.9.0
httptools==0.1.2
idna==3.4
inflect==5.6.2
Jinja2==2.11.3
kombu==4.6.11
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
more-itertools==9.1.0
oauthlib==3.2.2
packaging==23.0
passlib==1.7.4
pluggy==0.13.1
premailer==3.10.0
psycopg2-binary==2.9.5
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.6
PyJWT==2.6.0
pytest==5.4.3
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.5
python3-openid==3.2.0
pytz==2022.7.1
raven==6.10.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
sqlparse==0.4.3
starlette==0.16.0
tenacity==6.3.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.11.8
uvloop==0.17.0
vine==1.3.0
wcwidth==0.2.6
websockets==8.1
